Further optimizations to the lists
Moved "Want You Gone" into the same list template as what I moved "Still Alive" to in my last push. These arrays should make it a lot easier to transfer the songs to a JSON file later.
I would also like to note that the timings on Still Alive have not been fixed, I just copied what times were there into the lists.
I also added another quick optimization to the code accessing the lists on "Still Alive", it was a very small change in the code which most likely shouldn't effect the endpoint, but the code for it was really annoying me.